dashboard: one component for every kind of document the palette shows (#55828)
Consolidates the different types of document rows (files, scheduled jobs, and plain old documents) into one row type for easy re-use

GitOrigin-RevId: 3e3a1bbad9a993b7b388075ae6f4ead199a4ba7f
